+++ main_supcon.py
@@ -166,7 +166,7 @@ def set_loader(opt):
     normalize = transforms.Normalize(mean=mean, std=std)
 
     train_transform = transforms.Compose([
-        transforms.RandomResizedCrop(size=opt.img_size*0.875, scale=(0.2, 1.)),
+        transforms.RandomResizedCrop(size=int(opt.img_size*0.875), scale=(0.2, 1.)),
         transforms.RandomHorizontalFlip(),
         transforms.RandomApply([
             transforms.ColorJitter(0.4, 0.4, 0.4, 0.1)
